Декомпозировать FileBackedTaskManager: вынести логику работы с CSV в utils
